requirements.txt
groq==1.0.0
langchain_classic==1.0.1
langchain_community==0.4.1
langchain_core==1.2.14
langchain_experimental==0.4.1
langchain_groq==1.1.2
langchain_huggingface==1.2.0
langchain_text_splitters==1.1.1
nltk==3.9.2
numpy==2.4.2
python-dotenv==1.2.1
scikit_learn==1.8.0
streamlit==1.53.0
streamlit_pdf_viewer==0.0.27
